'02052019H1536'

+++ src/Controller/RespuestaController.php
@@ -2,41 +2,43 @@
 
 namespace App\Controller;
 
-use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use App\Entity\Respuesta;
-use App\Entity\Plantrabajo;
 use App\Form\RespuestaType;
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpFoundation\JsonResponse;
-use Symfony\Component\Form\Extension\Core\Type\DateTimeType;
-use Symfony\Component\Form\Extension\Core\Type\TextareaType;
-use App\Form\Transformer\DateTimetoStringTransformer;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use App\Entity\Fichero;
 use App\Entity\Actividad;
 
 /**
  * @Route("/respuesta")
  */
-class RespuestaController extends Controller
+class RespuestaController extends AbstractController
 {
     /**
      * @Route("/{id}/new", name="respuesta_new", methods="GET|POST")
      */
     public function new(Request $request, Actividad $actividad): Response
     {
+        if(!$request->isXmlHttpRequest())
+            throw $this->createAccessDeniedException();
+
+        $em = $this->getDoctrine()->getManager();
+        $existeRespuesta = $em->getRepository('App:Respuesta')->find($actividad)!=null;
+        if($existeRespuesta==true)
+            throw $this->createAccessDeniedException();
+
         $respuesta = new Respuesta();
         $respuesta->setId($actividad);
-//        $this->denyAccessUnlessGranted('NEW', $respuesta);
+        $this->denyAccessUnlessGranted('NEW', $respuesta);
         $form = $this->createForm(RespuestaType::class, $respuesta, array('action' => $this->generateUrl('respuesta_new', array('id' => $actividad->getId()))));
         $form->handleRequest($request);
         if ($form->isSubmitted())
             if ($form->isValid()) {
                 $ruta = $this->getParameter('storage_directory');
-                $em = $this->getDoctrine()->getManager();
+
                 foreach ($form->getData()->getFicheros() as $value) {
                     $value->setRespuesta($respuesta);
                     $value->subirArchivo($ruta);
@@ -46,11 +48,14 @@ class RespuestaController extends Controller
                 $em->persist($respuesta);
                 $em->flush();
 
-                return new JsonResponse(array('mensaje' => "La respuesta fue registrada satisfactoriamente"));
+                return new JsonResponse(array('mensaje' => "La respuesta fue registrada satisfactoriamente",
+                    'href'=>$this->generateUrl('respuesta_edit',['id'=>$respuesta->getId()->getId()])
+                ));
             } else {
                 $page = $this->renderView('respuesta/_form.html.twig', array(
                     'form' => $form->createView(),
                     'respuesta' => $respuesta,
+                    'existeRespuesta'=>$existeRespuesta
                 ));
                 return new JsonResponse(array('form' => $page, 'error' => true));
             }
@@ -58,15 +63,19 @@ class RespuestaController extends Controller
         return $this->render('respuesta/_new.html.twig', [
             'respuesta' => $respuesta,
             'form' => $form->createView(),
+            'existeRespuesta'=>$existeRespuesta
         ]);
     }
 
     /**
      * @Route("/{id}/show", name="respuesta_show", methods="GET")
      */
-    public function show(Respuesta $respuesta): Response
+    public function show(Request $request,Respuesta $respuesta): Response
     {
-        //$this->denyAccessUnlessGranted('VIEW', $respuesta);
+        if(!$request->isXmlHttpRequest())
+            throw $this->createAccessDeniedException();
+
+        $this->denyAccessUnlessGranted('VIEW', $respuesta);
         return $this->render('respuesta/show.html.twig', ['respuesta' => $respuesta]);
     }
 
@@ -75,9 +84,11 @@ class RespuestaController extends Controller
      */
     public function edit(Request $request, Respuesta $respuesta): Response
     {
-      //  $this->denyAccessUnlessGranted('EDIT', $respuesta);
+        if(!$request->isXmlHttpRequest())
+            throw $this->createAccessDeniedException();
+
+        $this->denyAccessUnlessGranted('EDIT', $respuesta);
         $form = $this->createForm(RespuestaType::class, $respuesta, array('action' => $this->generateUrl('respuesta_edit', array('id' => $respuesta->getId()->getId()))));
-        $ficherosIniciales = $respuesta->getFicheros()->toArray();
         $form->handleRequest($request);
         if ($form->isSubmitted())
             if ($form->isValid()) {
@@ -98,6 +109,7 @@ class RespuestaController extends Controller
                     'form' => $form->createView(),
                     'action' => 'Actualizar',
                     'form_id' => 'respuesta_edit',
+                    'existeRespuesta'=>true
                 ));
                 return new JsonResponse(array('form' => $page, 'error' => true));
             }
@@ -108,6 +120,7 @@ class RespuestaController extends Controller
             'action' => 'Actualizar',
             'form_id' => 'respuesta_edit',
             'form' => $form->createView(),
+            'existeRespuesta'=>true
         ]);
     }
 
@@ -118,11 +131,14 @@ class RespuestaController extends Controller
     public function delete(Request $request, Respuesta $respuesta): Response
     {
         if ($request->isXmlHttpRequest() && $this->isCsrfTokenValid('delete' . $respuesta->getId()->getId(), $request->query->get('_token'))) {
-            //  $this->denyAccessUnlessGranted('DELETE', $respuesta);
+            $this->denyAccessUnlessGranted('DELETE', $respuesta);
+            $actividad_id=$respuesta->getId()->getId();
             $em = $this->getDoctrine()->getManager();
             $em->remove($respuesta);
             $em->flush();
-            return new JsonResponse(array('mensaje' => "La respuesta fue eliminada satisfactoriamente"));
+            return new JsonResponse(array('mensaje' => "La respuesta fue eliminada satisfactoriamente",
+                'href'=>$this->generateUrl('respuesta_new',['id'=>$actividad_id])
+            ));
         }
 
         throw $this->createAccessDeniedException();
